+++ CHANGELOG.md
@@ -4,6 +4,24 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [0.5.0] - 2016-05-09
+### Added
+- Various example apps.
+- Warning message (`console.warn`) if `@Body` is used without `body-parser` package.
+- More validators (i.e. parametrized).
+
+### Changed
+- Validators are now factories.
+- Validators have new names (*breaking*).
+
+### Fixed
+- `npm test` command not working.
+- Removed *.spec files from coverage test.
+- Already parsed values are not parsed again.
+- Fixed this context in error handlers.
+
+
 ## [0.4.0] - 2016-04-22
 ### Added
 - HTTP head as route annotation.
@@ -19,6 +37,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Removed coveralls from `package.json` and moved its installation to Travis configuration
 
+
 ## [0.3.0] - 2016-04-20
 ### Added
 - `@Header` parameter description.
@@ -26,12 +45,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Parameter parsing with undefined values.
 
+
 ## [0.2.1] - 2016-04-20
 ### Added
 - First version of giuseppe.
 - Adding base code and travis-CI.
 
-[Unreleased]: https://github.com/smartive/giuseppe/compare/v0.4.0...develop
-[0.4.0]: https://github.com/smartive/giuseppe/compare/v0.4.0...v0.3.1
+[Unreleased]: https://github.com/smartive/giuseppe/compare/v0.5.0...develop
+[0.5.0]: https://github.com/smartive/giuseppe/compare/v0.4.0...v0.5.0
+[0.4.0]: https://github.com/smartive/giuseppe/compare/v0.3.1...v0.4.0
 [0.3.0]: https://github.com/smartive/giuseppe/compare/v0.2.0...v0.3.0
 [0.2.1]: https://github.com/smartive/giuseppe/tree/v0.2.1